pyproject.toml
[project]
name = "genai_processors"
description = "GenAI Processors Library"
readme = "README.pypi.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [{name = "Google DeepMind", email="noreply@google.com"}]
classifiers = [  # List of https://pypi.org/classifiers/
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = []

# pip dependencies of the project
# Installed locally with `pip install -e .`
dependencies = [
    "absl-py>=1.0.0",
    "dataclasses-json>=0.6.0",
    "google-genai>=1.16.0",
    "google-api-python-client>=0.6.0",
    "google-cloud-texttospeech>=2.27.0",
    "google-cloud-speech>=2.33.0",
    "httpx>=0.24.0",
    "jinja2>=3.0.0",
    "opencv-python>=2.0.0",
    "numpy>=2.0.0",
    "pdfrw>=0.4",
    "Pillow>=9.0.0",
    "termcolor>=3.0.0",
    "pypdfium2>=4.30.0",
    "xxhash>=3.0.0",
]

# `version` is automatically set by flit to use `genai_processors.__version__`
dynamic = ["version"]

[project.urls]
repository = "https://github.com/google-gemini/genai-processors"

[project.optional-dependencies]
# Dependencies of processors in contrib. To avoid dependency bloat we do not
# include them in the main dependency list. But they can be installed with
# `pip install -e .[contrib]`
contrib = [
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.7",
]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install -e .[dev]`
dev = [
    "pytest",
    "pytest-xdist",
    "pylint>=2.6.0",
    "pyink",
    "flake8",
]

[tool.pyink]
# Formatting configuration to follow Google style-guide
line-length = 80
preview = true
pyink-indentation = 2
pyink-use-majority-quotes = true

[build-system]
# Build system specify which backend is used to build/install the project (flit,
# poetry, setuptools,...). All backends are supported by `pip install`
requires = ["flit_core >=3.8,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "genai_processors"
dir = "."
[tool.flit.sdist]
exclude = ["tests/", "tests/*"]

